Fix duplicated file.
